drivers: input: misc: bma023: set correct ioctl pointer

+++ drivers/input/misc/bma023.c
@@ -836,7 +836,7 @@ static int bma023_close(struct inode *inode, struct file *file)
 	return 0;
 }
 
-static int bma023_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
+static long bma023_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 {
 	struct bma023_data *bma023 = file->private_data;
 	int try;
